// src/app/model/zwb-aggregate-corp.ts
export interface ZWBAggregateCorp {
  corporation: string;
  kills: number;
  losses: number;
  kdratio: number;
  kdefficiency: number;
  iskwon: number;
  isklost: number;
  netisk: number;
  iskefficiency: number;
  numactive: number;
  avgperkill: number;
  iskperactive: number;
  netiskperactive: number;
  iskperavgonkill: number;
  netiskperavgonkill: number;
}